+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## unreleased
+* [ENHANCEMENT] Add `statsd_exporter_metrics_current` gauge for the current number of active metric vectors by type.
+
 ## 0.30.0 / 2026-05-28
 * [CHANGE] Remove the Dockerfile `HEALTHCHECK` from published container images ([#671](https://github.com/prometheus/statsd_exporter/pull/671))
 * [ENHANCEMENT] Add a distroless container image variant ([#703](https://github.com/prometheus/statsd_exporter/pull/703))